#401 refactored gui package for clarity
